# docs/specs/swagger.yaml
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "twitter-challange-espenia",
    "description" : "@project.description@",
    "version" : "@project.version@"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Scope local"
  } ],
  "paths" : {
    "/api/users" : {
      "post" : {
        "tags" : [ "user-controller" ],
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tweets" : {
      "post" : {
        "tags" : [ "tweet-controller" ],
        "operationId" : "createTweet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TweetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TweetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/follows/{followerId}" : {
      "post" : {
        "tags" : [ "follow-controller" ],
        "operationId" : "createFollow",
        "parameters" : [ {
          "name" : "followerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FollowRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FollowResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "getUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "user-controller" ],
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "user-controller" ],
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ping" : {
      "get" : {
        "tags" : [ "ping-controller" ],
        "operationId" : "ping",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/username/{username}" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "operationId" : "getUserByUsername",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/follows/following/{userId}" : {
      "get" : {
        "tags" : [ "follow-controller" ],
        "operationId" : "getFollowing",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FollowResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/follows/followers/{userId}" : {
      "get" : {
        "tags" : [ "follow-controller" ],
        "operationId" : "getFollowers",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FollowResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/follows/{followerId}/{followedId}" : {
      "delete" : {
        "tags" : [ "follow-controller" ],
        "operationId" : "deleteFollow",
        "parameters" : [ {
          "name" : "followerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "followedId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "UserRequest" : {
        "required" : [ "bio", "display_name", "email", "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "maxLength" : 50,
            "minLength" : 3,
            "type" : "string"
          },
          "display_name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "maxLength" : 100,
            "minLength" : 8,
            "type" : "string"
          },
          "bio" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "display_name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          }
        }
      },
      "TweetRequest" : {
        "required" : [ "content", "user_id" ],
        "type" : "object",
        "properties" : {
          "user_id" : {
            "type" : "string"
          },
          "content" : {
            "maxLength" : 280,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "TweetResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "user_id" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "like_count" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "FollowRequest" : {
        "required" : [ "followed_id" ],
        "type" : "object",
        "properties" : {
          "followed_id" : {
            "type" : "string"
          }
        }
      },
      "FollowResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "follower_id" : {
            "type" : "string"
          },
          "followed_id" : {
            "type" : "string"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UserUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "display_name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 100,
            "minLength" : 8,
            "type" : "string"
          },
          "bio" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        }
      }
    }
  }
}